% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_log_plastic.R
\name{analyze_log_plastic}
\alias{analyze_log_plastic}
\title{Analyze the linear regression of log(Total Plastic Collected) on Mean Annual Temperature.}
\usage{
analyze_log_plastic()
}
\value{
A gt table with simple linear regression coefficient estimates, standard errors, t-tests, p-values, and model fit statistics.
}
\description{
Analyze the linear regression of log(Total Plastic Collected) on Mean Annual Temperature.
}
