% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_plastic.R
\name{analyze_plastic}
\alias{analyze_plastic}
\title{Linear regression of Total Plastic Collected on Mean Annual Temperature}
\usage{
analyze_plastic(log_transform = TRUE)
}
\value{
A gt table with simple linear regression coefficient estimates,
standard errors, t-tests, p-values, and model fit statistics.
}
\description{
Fits a simple linear regression model using mean annual temperature to
predict total plastic collected (with an optional a log transformation)
and formats the coefficient estimates and model fit statistics as a gt table.
}
