% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nigeria.R
\name{clean_nigeria}
\alias{clean_nigeria}
\title{Clean the Nigeria plastic pollution dataset}
\usage{
clean_nigeria()
}
\value{
A tibble with neighborhood, plastic type, and count cleaned.
}
\description{
Clean the Nigeria plastic pollution dataset
}
