% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_nigeria_data}
\alias{load_nigeria_data}
\title{Load Nigeria Dataset}
\usage{
load_nigeria_data()
}
\value{
A tibble containing plastic waste collection data from the Break Free From Plastic movement.
}
\description{
Downloads outside data from the Earth Challenge 2020 for beach-cleanups in Nigeria.
}
