% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_plastic_data}
\alias{load_plastic_data}
\title{Load Plastic Waste Dataset}
\usage{
load_plastic_data()
}
\value{
A tibble containing plastic waste collection data from the Break Free
From Plastic movement.
}
\description{
Loads a cleaned plastic waste dataset containing plastic types, collection
counts, and mean annual temperatures for different countries used to analyze
relationships between plastic collection totals and temperatures.
}
