% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_plot.R
\name{plot_temp_plastic_hex}
\alias{plot_temp_plastic_hex}
\title{Plot Temperature and Plastic Collected}
\usage{
plot_temp_plastic_hex()
}
\value{
A ggplot object.
}
\description{
Creates a hexbin plot comparing mean annual temperature and total plastic collected.
}
