% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_by_cont_plot.R
\name{temp_by_cont_plot}
\alias{temp_by_cont_plot}
\title{Plot Temperature and Plastic Collected, by Continent}
\usage{
temp_by_cont_plot()
}
\value{
A ggplot object.
}
\description{
Creates a scatterplot of mean annual temperature and total plastic collected, colored by temperature.
}
