% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_items.R
\name{top_plastic_items}
\alias{top_plastic_items}
\title{Get the top n plastic types & counts for each type in the neighborhood specified in Nigeria.}
\usage{
top_plastic_items(neighborhood, n)
}
\arguments{
\item{neighborhood}{A character string of a neighborhood in Nigeria in the data.}

\item{n}{A numeric whole value less than or equal to 10}
}
\value{
A tibble of top plastic types with counts.
}
\description{
Get the top n plastic types & counts for each type in the neighborhood specified in Nigeria.
}
