% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_neighborhood}
\alias{validate_neighborhood}
\title{Validate Nigeria neighborhood name}
\usage{
validate_neighborhood(neighborhood)
}
\arguments{
\item{neighborhood}{A character string of a neighborhood in Nigeria}
}
\value{
Returns \code{TRUE} when the input is valid
}
\description{
Internal helper that checks whether Nigerian neighborhood name is valid
}
