% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_nigeria_data}
\alias{load_nigeria_data}
\title{Load Nigeria Dataset}
\usage{
load_nigeria_data()
}
\value{
A tibble containing plastic waste collection data from the Break Free
From Plastic movement.
}
\description{
Loads beach cleanup data collected in Nigeria through the Earth Challenge
2020 initiative, including neighborhood information and plastic collection
counts for different types of plastic.
}
