% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_plot.R
\name{plot_temp_plastic_hex}
\alias{plot_temp_plastic_hex}
\title{Plot Temperature and Plastic Collected}
\usage{
plot_temp_plastic_hex()
}
\value{
A ggplot object.
}
\description{
Shows the relationship between mean annual temperature and total
plastic collected using a hexbin plot of plastic counts, with a fitted linear
trend line and label of the country with the highest collection total.
}
