% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_summary.R
\name{temp_summary}
\alias{temp_summary}
\title{Create total plastic summary tables}
\usage{
temp_summary(group = c("continent", "temp_bin"))
}
\arguments{
\item{group}{Choose "continent" or "temp_bin".}
}
\value{
A gt table.
}
\description{
Summarizes total plastic collected by either continent or annual temperature
bins, calculating mean, median, and standard deviation values and formatting
the results in gt table with colored summary columns.
}
